refactor: move files over to the login module
Move some of the details screen over to the login module
