+++ login/src/main/java/com/chesire/malime/login/details/LoginStatus.kt
@@ -1,5 +1,8 @@
-package com.chesire.malime.flow.login.details
+package com.chesire.malime.login.details
 
+/**
+ * Provides the status of a login request.
+ */
 enum class LoginStatus {
     EmptyUsername,
     EmptyPassword,
